big pixi refactor
